Add ChoroplethLegend and styles component
Added ChoroplethLegend and styles component and removed legend from ChoroplethToggles.
