+++ src/App.js
@@ -21,6 +21,7 @@ import {
 } from './actions/mapActions';
 import MobileMessage from './components/mobileMessage';
 import ChoroplethToggles from './components/ChoroplethToggles';
+import ChoroplethLegend from './components/ChoroplethLegend';
 import DropdownSelectionStyles from './components/styles/DropdownSelectionStyles';
 import DirectionPad from './components/DirectionPad';
 import QuestionBox from './components/quizBox/questionBox';
@@ -191,6 +192,8 @@ class App extends Component {
         <InfoTab />
 
         <DirectionPad />
+
+        <ChoroplethLegend />
 
         <Button
           icon="sidebar"
