+++ src/components/styles/ChoroplethTogglesStyles.js
@@ -44,35 +44,6 @@ const ChoroplethTogglesStyles = styled.div`
     font-weight: 600;
     /* font-size: .8em */
   }
-
-  .chorolegend {
-    position: absolute;
-    bottom: 6rem;
-    left: 1em;
-    padding: 1em;
-    font-size: ${props => (props.isMobile ? '0.5em' : '1em')};
-    background: rgba(0, 0, 0, 0.5);
-    color: white;
-    border-radius: 1rem;
-    letter-spacing: 1px;
-
-    .legendTitle {
-      margin-bottom: 0.7em;
-      font-weight: 600;
-    }
-
-    .legendItem {
-      display: flex;
-      flex-direction: row;
-      line-height: 1.5em;
-      /* line-height: ${props => (props.isMobile ? '1.4em' : '2em')}; */
-    }
-
-    .legendColor {
-      margin-right: 5px;
-      width: 2em;
-    }
-  }
 `;
 
 export default ChoroplethTogglesStyles;
